+++ dao/src/main/java/com/company/webapp/order/dao/OrderRowMapper.java
@@ -11,10 +11,10 @@ public class OrderRowMapper implements RowMapper<Order> {
     @Override
     public Order mapRow(ResultSet resultSet, int rowNumber) throws SQLException {
         Order order = new Order();
-        order.setOrderId(resultSet.getInt("ORDER_ID"));
+        order.setOrderId(resultSet.getLong("ORDER_ID"));
         order.setName(resultSet.getString("ORDER_NAME"));
         order.setCreateDate(resultSet.getDate("ORDER_ADDDATE"));
-        order.setUserId(resultSet.getInt("USER_ID"));
+        order.setUserId(resultSet.getLong("USER_ID"));
         return order;
     }
 }
